Split unit tests and integration tests (#337)
* Split unit tests and integration tests

Integration tests requires a consul agent to be running with correct configuration.
When developing a feature, it would be great to be able to launch unit tests first.
Hence, we here purpose to:
 - have unit test classes in src/test with Test suffix
 - have integration test classes in src/itest with ITest suffix

* Fix agent test
