+++ src/itest/java/com/orbitz/consul/PreparedQueryITest.java
@@ -11,7 +11,7 @@ import org.junit.Test;
 
 import java.util.UUID;
 
-public class PreparedQueryTests {
+public class PreparedQueryITest {
 
     @Test
     @Ignore
