+++ src/test/java/com/orbitz/consul/util/SecondsSerDeTest.java
@@ -12,7 +12,7 @@ import java.util.Random;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-public class SecondsSerDeTests {
+public class SecondsSerDeTest {
 
     static class Item {
         @JsonSerialize(using = SecondsSerializer.class)
